Add code and fix layout in modules  attendance
